// bootstrap/compiler/src/generated/graph-capabilities.contracts.ts
export type GraphDeriveAdapter = {
	derive: (input: ModelTypes.GraphDeriveInput) => Promise<ModelTypes.DynamicDerivationResult>;
};
// ...
export type GraphProvideAdapter = {
	provide: (input: ModelTypes.GraphProviderInput) => Promise<ModelTypes.GraphProviderResult>;
};
// ...
export type GraphSelectorsMaterializeAdapter = {
	materialize: (input: ModelTypes.GraphSelectorsMaterializeInput) => Promise<ModelTypes.GraphSelectorsMaterializeResult>;
};